DLS-11614: Tidied up and re-arranged model objects and json formatters
