+++ app/uk/gov/hmrc/cgtpropertydisposalsstubs/controllers/SubscriptionProfiles.scala
@@ -19,13 +19,12 @@ package uk.gov.hmrc.cgtpropertydisposalsstubs.controllers
 import cats.instances.either.*
 import cats.syntax.either.*
 import play.api.mvc.Result
-import uk.gov.hmrc.cgtpropertydisposalsstubs.controllers.BusinessPartnerRecordController.DesBusinessPartnerRecord
-import uk.gov.hmrc.cgtpropertydisposalsstubs.controllers.BusinessPartnerRecordController.DesBusinessPartnerRecord.{DesContactDetails, DesIndividual, DesOrganisation}
-import play.api.mvc.Results.*
+import play.api.mvc.Results.{BadRequest, Conflict, Forbidden, InternalServerError, NotFound, ServiceUnavailable}
 import uk.gov.hmrc.cgtpropertydisposalsstubs.controllers.SubscriptionController.SubscriptionResponse
+import uk.gov.hmrc.cgtpropertydisposalsstubs.models.DesBusinessPartnerRecord.DesContactDetails
+import uk.gov.hmrc.cgtpropertydisposalsstubs.models.{DesAddressDetails, DesBusinessPartnerRecord, DesIndividual, DesOrganisation, NINO, SAUTR, SapNumber, SubscriptionStatusResponse, TRN}
 import uk.gov.hmrc.cgtpropertydisposalsstubs.models.DesErrorResponse.desErrorResponseJson
 import uk.gov.hmrc.cgtpropertydisposalsstubs.models.SubscriptionStatusResponse.SubscriptionStatus
-import uk.gov.hmrc.cgtpropertydisposalsstubs.models._
 
 case class Profile(
   predicate: Either[Either[TRN, SAUTR], NINO] => Boolean,
