+++ app/uk/gov/hmrc/cgtpropertydisposalsstubs/models/BprRequest.scala
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2025 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package uk.gov.hmrc.cgtpropertydisposalsstubs.models
+
+import play.api.libs.json.{Json, Reads}
+
+case class BprRequest(
+  regime: String,
+  requiresNameMatch: Boolean,
+  isAnAgent: Boolean,
+  individual: Option[DesIndividual],
+  organisation: Option[DesOrganisation]
+)
+
+object BprRequest {
+  implicit val bprRequestReads: Reads[BprRequest] = Json.reads[BprRequest]
+}
